Renamed MyCodeCard to MyQrCodeCard
