+++ ui/src/lib/components/contacts/MyQrCodeCard.svelte
